% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bericht.R
\name{bericht}
\alias{bericht}
\title{Bericht PDF Template}
\usage{
bericht(
  toc = FALSE,
  toc_depth = 2,
  number_sections = FALSE,
  fig_width = 6.5,
  fig_height = 4.5,
  fig_crop = TRUE,
  fig_caption = TRUE,
  df_print = "kable",
  highlight = "default",
  keep_tex = FALSE,
  pandoc_args = NULL,
  dev = "cairo_pdf",
  latex_engine = "xelatex",
  citation_package = "biblatex",
  ...
)
}
\arguments{
\item{number_sections}{Whether to number section headers: if \code{TRUE},
figure/table numbers will be of the form \code{X.i}, where \code{X} is the
current first-level section number, and \code{i} is an incremental number
(the i-th figure/table); if \code{FALSE}, figures/tables will be numbered
sequentially in the document from 1, 2, ..., and you cannot cross-reference
section headers in this case.}

\item{fig_caption}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{pandoc_args}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}

\item{...}{Arguments to be passed to a
specific output format function. For a function \code{foo2()}, its
arguments are passed to \code{foo()}, e.g. \code{...} of
\code{html_document2()} are passed to \code{rmarkdown::html_document()}.}
}
\description{
Bericht PDF Template
}
