% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterthesis.R
\name{masterthesis}
\alias{masterthesis}
\title{Master thesis bookdown PDF template}
\usage{
masterthesis(...)
}
\arguments{
\item{...}{Other arguments to be passed to \code{base_format}.}
}
\description{
See \code{bookdown::pdf_book()} for options.
}
